--- Source/P_MA/Private/MAGameplayWidget.cpp
@@ -1,20 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "MAGameplayWidget.h"
-#include "AbilitySystemBlueprintLibrary.h"
-#include "AbilitySystemComponent.h"
-#include "Widget/MAValueGauge.h"
-#include "GAS/MAAttributeSet.h"
-
-void UMAGameplayWidget::NativeConstruct()
-{
-	Super::NativeConstruct();
-
-	OwnerAbilitySystemComponent = UAbilitySystemBlueprintLibrary::GetAbilitySystemComponent(GetOwningPlayerPawn());
-	if (OwnerAbilitySystemComponent)
-	{
-		HealthBar->SetAndBoundToGameplayAttribute(OwnerAbilitySystemComponent, UMAAttributeSet::GetHealthAttribute(), UMAAttributeSet::GetMaxHealthAttribute());
-	}
-}
-
