--- Source/P_MA/Public/MAGameplayWidget.h
@@ -1,24 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Blueprint/UserWidget.h"
-#include "MAGameplayWidget.generated.h"
-
-/**
- *
- */
-UCLASS()
-class UMAGameplayWidget : public UUserWidget
-{
-	GENERATED_BODY()
-public:
-	virtual void NativeConstruct() override;
-private:
-	UPROPERTY(meta = (BindWidget))
-	class UMAValueGauge* HealthBar;
-
-	UPROPERTY()
-	class UAbilitySystemComponent* OwnerAbilitySystemComponent;
-};
