--- Source/P_MA/Public/Widget/MAHUD.h
@@ -1,44 +1,0 @@
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Blueprint/UserWidget.h"
-#include "MAHUD.generated.h"
-
-class UHealthBarWidget;
-class UMASkillSlotWidget;
-class UHorizontalBox;
-class UMAPassiveSlotWidget;
-
-UCLASS()
-class P_MA_API UMAHUD : public UUserWidget
-{
-    GENERATED_BODY()
-
-public:
-    virtual void NativeConstruct() override;
-
-    void UpdateHealth(float CurrentHealth, float MaxHealth);
-
-protected:
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "UI")
-    TSubclassOf<UHealthBarWidget> HealthBarWidgetClass;
-
-    UPROPERTY(meta = (BindWidget))
-    UHealthBarWidget* HealthBarWidget;
-
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "UI")
-    TSubclassOf<UMASkillSlotWidget> SkillSlotWidgetClass;
-
-    UPROPERTY(meta = (BindWidget))
-    UHorizontalBox* HorizontalBox_SkillSlots;
-
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "UI")
-    TSubclassOf<UMAPassiveSlotWidget> PassiveSlotWidgetClass;
-
-    UPROPERTY(meta = (BindWidget))
-    UHorizontalBox* HorizontalBox_PassiveSlots;
-
-private:
-    void CreateSkillSlots(int32 NumSlots);
-    void CreatePassiveSlots(int32 NumSlots);
-};
